make conversation read after received "carbon sent"
it means, that user send response in conversation from another resource

+++ app/src/main/java/org/tigase/messenger/jaxmpp/android/chat/MarkAsRead.java
@@ -20,7 +20,7 @@ public class MarkAsRead {
 		this.context = context.getApplicationContext();
 	}
 
-	private void intMarkAsRead(final Uri u, final int chatId, final BareJID account, final JID jid) {
+	private void intMarkAsRead(final Uri u, final long chatId, final BareJID account, final JID jid) {
 		(new AsyncTask<Void, Void, Void>() {
 			@Override
 			protected Void doInBackground(Void... params) {
@@ -51,11 +51,11 @@ public class MarkAsRead {
 		}).execute();
 	}
 
-	public void markChatAsRead(final int chatId, final BareJID account, final JID jid) {
+	public void markChatAsRead(final long chatId, final BareJID account, final JID jid) {
 		intMarkAsRead(ChatProvider.CHAT_HISTORY_URI, chatId, account, jid);
 	}
 
-	public void markGroupchatAsRead(int openChatId, BareJID account, JID jid) {
+	public void markGroupchatAsRead(long openChatId, BareJID account, JID jid) {
 		intMarkAsRead(ChatProvider.MUC_HISTORY_URI, openChatId, account, jid);
 	}
 }
